웹폴더 즐겨찾기 컨트롤러 수정 (userId, serverName만 넘겨주도록)

+++ src/main/java/egovframework/ezEKP/ezWebFolder/web/EzWebFolderGWController_m.java
@@ -506,12 +506,9 @@ public class EzWebFolderGWController_m {
 	 */
 	@RequestMapping(value = "/rest/ezwebfolder/users/{userId}/favorites", method = RequestMethod.GET, produces = "application/json;charset=utf-8")
 	public JSONObject getFavoriteList(@PathVariable String userId, HttpServletRequest request) {
-		logger.debug("REST | getUserFavorites started.");
-
-		String offset = request.getParameter("offset");
-		String primaryLang = orElse(request.getParameter("primary"), "1");
-		int tenantId = Integer.parseInt(orElse(request.getParameter("tenantId"), "0"));
-
+		logger.debug("G/W WEBFOLDER [GET /rest/ezwebfolder/users/{userId}/favorites] started.");
+		
+		String serverName = request.getHeader("x-user-host");
 		int startIndex = Integer.parseInt(orElse(request.getParameter("startIndex"), "0"));
 		int listCount = Integer.parseInt(orElse(request.getParameter("listCount"), "0"));
 		
@@ -523,25 +520,41 @@ public class EzWebFolderGWController_m {
 		searchInfo.setSearchStartDate(orElse(request.getParameter("searchStartDate"), ""));
 		searchInfo.setSearchEndDate(orElse(request.getParameter("searchEndDate"), ""));
 		
-		// list count
-		int userListCount = ezWebFolderService_y.getUsrListCount(tenantId, userId);
+		JSONObject result = new JSONObject();
 		
-		if (listCount == 0) {
-			listCount = userListCount;
+		if (containsNull(serverName, userId)) {
+			result.put("status", "error");
+			result.put("code", 1);
+			
+			logger.debug("G/W WEBFOLDER parameter error. [GET /rest/ezwebfolder/users/{userId}/favorites] ended.");
+			return result;
 		}
 		
-		ezWebFolderService_y.insertEnv(userId, tenantId, listCount);
-		
-		JSONObject result = new JSONObject();
-		JSONObject data = new JSONObject();
-
 		try {
-			List<FavoriteVO> favoriteFiles = ezWebFolderService_m.getFavorites(userId, primaryLang, offset, tenantId, searchInfo, startIndex, listCount);
+			MCommonVO userInfo = mOptionService.commonInfoWeb(serverName, userId);
+			
+			int tenantId = userInfo.getTenantId();
+			String primary = userInfo.getPrimary();
+			String offset = userInfo.getOffSet();
+			
+			// setup list count
+			{
+				// list count
+				int userListCount = ezWebFolderService_y.getUsrListCount(tenantId, userId);
+				
+				if (listCount == 0) {
+					listCount = userListCount;
+				}
+				
+				ezWebFolderService_y.insertEnv(userId, tenantId, listCount);
+			}
+			
+			List<FavoriteVO> favoriteFiles = ezWebFolderService_m.getFavorites(userId, primary, offset, tenantId, searchInfo, startIndex, listCount);
 			String targetPath;
 			
 			for (FavoriteVO favoriteFile : favoriteFiles) {
 				targetPath = favoriteFile.getTargetPath().substring(1);
-				targetPath = getFolderPath(targetPath.split("\\|"), offset, primaryLang, tenantId);
+				targetPath = getFolderPath(targetPath.split("\\|"), offset, primary, tenantId);
 				
 				// is folder
 				if ("D".equalsIgnoreCase(favoriteFile.getTargetType())) {
@@ -555,6 +568,7 @@ public class EzWebFolderGWController_m {
 			}
 			
 			Map<String, Integer> favoriteCountMap = ezWebFolderService_m.getFavoriteCount(userId, offset, tenantId, searchInfo);
+			JSONObject data = new JSONObject();
 			
 			data.put("totalCount", favoriteCountMap.get("totalCount"));
 			data.put("folderCount", favoriteCountMap.get("folderCount"));
@@ -569,12 +583,11 @@ public class EzWebFolderGWController_m {
 		} catch (Exception e) {
 			e.printStackTrace();
 			result.put("status", "error");
-			result.put("code", 1);
-			result.put("data", "");
+			result.put("code", 2);
 		}
 
 		logger.debug(String.format("result: %s", result.toJSONString()));
-		logger.debug("REST | getUserFavorites ended.");
+		logger.debug("G/W WEBFOLDER [GET /rest/ezwebfolder/users/{userId}/favorites] ended.");
 		return result;
 	}
 	
@@ -593,27 +606,39 @@ public class EzWebFolderGWController_m {
 	 */
 	@RequestMapping(value = "/rest/ezwebfolder/users/{userId}/favorite", method = RequestMethod.POST, produces = "application/json;charset=utf-8")
 	public JSONObject addFavorite(@PathVariable String userId, HttpServletRequest request) {
-		logger.debug("REST | addUserFavorite started.");
+		logger.debug("G/W WEBFOLDER [POST /rest/ezwebfolder/users/{userId}/favorite] started.");
 		
+		String serverName = request.getHeader("x-user-host");
 		String targetId = request.getParameter("targetId");
 		String targetType = request.getParameter("targetType");
-		String offset = request.getParameter("offset");
-		int tenantId = Integer.parseInt(orElse(request.getParameter("tenantId"), "0"));
-
-		SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-		String createDate = commonUtil.getDateStringInUTC(format.format(new Date()), offset, true);
-
+		
 		JSONObject result = new JSONObject();
 
+		
+		if (containsNull(serverName, userId)) {
+			result.put("status", "error");
+			result.put("code", 1);
+			
+			logger.debug("G/W WEBFOLDER parameter error. [POST /rest/ezwebfolder/users/{userId}/favorite] ended.");
+			return result;
+		}
+
 		try {
+			MCommonVO userInfo = mOptionService.commonInfoWeb(serverName, userId);
+			int tenantId = userInfo.getTenantId();
+			String offset = userInfo.getOffSet();
+			
+			SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+			String createDate = commonUtil.getDateStringInUTC(format.format(new Date()), offset, true);
+			
 			if(ezWebFolderService_m.isExistsFavorite(userId, targetId, targetType, tenantId)) {
-				result.put("code", 1);
+				result.put("status", "error");
+				result.put("code", 2);
 			} else {
 				ezWebFolderService_m.addFavorite(userId, targetId, targetType, createDate, tenantId);
+				result.put("status", "ok");
 				result.put("code", 0);
 			}
-			
-			result.put("status", "ok");
 		} catch (Exception e) {
 			e.printStackTrace();
 			result.put("status", "error");
@@ -621,7 +646,7 @@ public class EzWebFolderGWController_m {
 		}
 
 		logger.debug(String.format("result: %s", result.toJSONString()));
-		logger.debug("REST | addUserFavorite ended.");
+		logger.debug("G/W WEBFOLDER [POST /rest/ezwebfolder/users/{userId}/favorite] ended.");
 		return result;
 	}
 	
@@ -640,23 +665,34 @@ public class EzWebFolderGWController_m {
 	 */
 	@RequestMapping(value = "/rest/ezwebfolder/users/{userId}/favorite", method = RequestMethod.DELETE, produces = "application/json;charset=utf-8")
 	public JSONObject deleteFavorite(@PathVariable String userId, HttpServletRequest request, @RequestBody JSONObject jsonObject) {
-		logger.debug("REST | deleteUserFavorite started.");
+		logger.debug("G/W WEBFOLDER [DELETE /rest/ezwebfolder/users/{userId}/favorite] started.");
 
+		String serverName = request.getHeader("x-user-host");
 		String targetId = (String) jsonObject.get("targetId");
 		String targetType = (String) jsonObject.get("targetType");
-		int tenantId = (Integer) jsonObject.get("tenantId");
 
 		JSONObject result = new JSONObject();
 
+		if (containsNull(serverName, userId)) {
+			result.put("status", "error");
+			result.put("code", 1);
+			
+			logger.debug("G/W WEBFOLDER parameter error. [DELETE /rest/ezwebfolder/users/{userId}/favorite] ended.");
+			return result;
+		}
+		
 		try {
+			MCommonVO userInfo = mOptionService.commonInfoWeb(serverName, userId);
+			int tenantId = userInfo.getTenantId();
+			
 			if(ezWebFolderService_m.isExistsFavorite(userId, targetId, targetType, tenantId)) {
 				ezWebFolderService_m.deleteFavorite(userId, targetId, targetType, tenantId);
+				result.put("status", "ok");
 				result.put("code", 0);
 			} else {
-				result.put("code", 1);
+				result.put("status", "error");
+				result.put("code", 2);
 			}
-			
-			result.put("status", "ok");
 		} catch (Exception e) {
 			e.printStackTrace();
 			result.put("status", "error");
@@ -664,7 +700,7 @@ public class EzWebFolderGWController_m {
 		}
 
 		logger.debug(String.format("result: %s", result.toJSONString()));
-		logger.debug("REST | deleteUserFavorite ended.");
+		logger.debug("G/W WEBFOLDER [DELETE /rest/ezwebfolder/users/{userId}/favorite] ended.");
 		return result;
 	}
 	
@@ -948,6 +984,21 @@ public class EzWebFolderGWController_m {
 		}
 		
 		return value != null ? value : other;
+	}
+	
+	private boolean containsNull(Object... elements) {
+		for(Object e : elements) {
+			if (e == null) {
+				return true;
+			}
+			
+			// string is ""
+			if (e.toString().isEmpty()) {
+				return true;
+			}
+		}
+		
+		return false;
 	}
 	
 	private String getFolderPath(String[] paths, String offset, String primaryLang, int tenantId) throws Exception {
